+++ routes/api/contacts-router.js
@@ -1,84 +1,15 @@
 import express from "express";
-import Joi from "joi";
-import contactsService from "../../models/contacts.js";
-import { HttpError } from "../../helpers/index.js";
-import {
-  nameSchema,
-  emailSchema,
-  phoneSchema,
-} from "../../helpers/validation-schemas.js";
-
+import ctrl from "../../controllers/contacts-controllers.js";
 const contactsRouter = express.Router();
 
-const contactAddSchema = Joi.object({
-  name: nameSchema,
-  email: emailSchema,
-  phone: phoneSchema,
-});
+contactsRouter.get("/", ctrl.getAllContacts);
 
-contactsRouter.get("/", async (req, res, next) => {
-  try {
-    const result = await contactsService.listContacts();
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+contactsRouter.get("/:id", ctrl.getById);
 
-contactsRouter.get("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await contactsService.getContactById(id);
-    if (!result) {
-      throw HttpError(404, `Contact with id=${id} not found`);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+contactsRouter.post("/", ctrl.addNewContact);
 
-contactsRouter.post("/", async (req, res, next) => {
-  try {
-    const { error } = contactAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const result = await contactsService.addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+contactsRouter.put("/:id", ctrl.updateContact);
 
-contactsRouter.put("/:id", async (req, res, next) => {
-  try {
-    const { error } = contactAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const { id } = req.params;
-    const result = await contactsService.updateContactById(id, req.body);
-    if (!result) {
-      throw HttpError(404, `Contact with id=${id} not found`);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-contactsRouter.delete("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await contactsService.removeContact(id);
-    if (!result) {
-      throw HttpError(404, `Contact with id=${id} not found`);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+contactsRouter.delete("/:id", ctrl.deleteContact);
 
 export default contactsRouter;
